fix: resolve Codacy quality and security issues
- Refactored encounters/routes.ts into modular files to meet 450-line limit
- Fixed XSS security vulnerabilities by implementing safe SSE data formatting
- Reduced cyclomatic complexity in checkRedisConnection function
- Simplified test setup by eliminating code duplication in Prisma mocks
- Refactored EncounterService methods to reduce complexity and line count
- Fixed markdown formatting issues in CLAUDE.md
- Split large route file into focused modules: basic-routes, combat-routes, participant-routes, stream-routes
- Added shared utilities module for common dependencies and helpers
