+++ packages/server/src/services/EncounterService.ts
@@ -113,18 +113,9 @@ export class EncounterService {
   }
 
   /**
-   * Update encounter basic information
+   * Validate encounter update data
    */
-  async updateEncounter(
-    encounterId: string,
-    userId: string,
-    data: {
-      name?: string;
-      description?: string;
-      status?: EncounterStatus;
-    }
-  ): Promise<EncounterWithDetails> {
-    // Validation
+  private validateEncounterUpdateData(data: { name?: string; description?: string; status?: EncounterStatus }): void {
     if (data.name !== undefined) {
       if (!data.name || data.name.trim().length === 0) {
         throw new Error('Encounter name is required');
@@ -133,8 +124,12 @@ export class EncounterService {
         throw new Error('Encounter name must be 100 characters or less');
       }
     }
+  }
 
-    // Verify ownership
+  /**
+   * Verify encounter ownership
+   */
+  private async verifyEncounterOwnership(encounterId: string, userId: string): Promise<void> {
     const encounter = await this.prisma.encounter.findUnique({
       where: { id: encounterId },
       select: { userId: true },
@@ -147,8 +142,14 @@ export class EncounterService {
     if (encounter.userId !== userId) {
       throw new Error('Not authorized to modify this encounter');
     }
+  }
 
+  /**
+   * Build update data object from provided fields
+   */
+  private buildUpdateData(data: { name?: string; description?: string; status?: EncounterStatus }): any {
     const updateData: any = {};
+    
     if (data.name !== undefined) {
       updateData.name = data.name.trim();
     }
@@ -158,6 +159,25 @@ export class EncounterService {
     if (data.status !== undefined) {
       updateData.status = data.status;
     }
+    
+    return updateData;
+  }
+
+  /**
+   * Update encounter basic information
+   */
+  async updateEncounter(
+    encounterId: string,
+    userId: string,
+    data: {
+      name?: string;
+      description?: string;
+      status?: EncounterStatus;
+    }
+  ): Promise<EncounterWithDetails> {
+    this.validateEncounterUpdateData(data);
+    await this.verifyEncounterOwnership(encounterId, userId);
+    const updateData = this.buildUpdateData(data);
 
     return this.prisma.encounter.update({
       where: { id: encounterId },
@@ -205,21 +225,8 @@ export class EncounterService {
     userId: string,
     participantData: ParticipantCreateData
   ): Promise<EncounterWithDetails> {
-    // Verify encounter ownership
-    const encounter = await this.prisma.encounter.findUnique({
-      where: { id: encounterId },
-      select: { userId: true },
-    });
+    await this.verifyEncounterOwnership(encounterId, userId);
 
-    if (!encounter) {
-      throw new Error('Encounter not found');
-    }
-
-    if (encounter.userId !== userId) {
-      throw new Error('Not authorized to modify this encounter');
-    }
-
-    // Add participant
     await this.prisma.participant.create({
       data: {
         encounterId,
@@ -238,7 +245,6 @@ export class EncounterService {
       },
     });
 
-    // Return updated encounter
     return this.getEncounterById(encounterId) as Promise<EncounterWithDetails>;
   }
